Begin a branch of the website content for conversion to use as ikiwiki source, from which the "real" website will then be generated.
Most of the html pages in the base website will, initially, be left
unchanged to be used as rawhtml by ikiwiki, however the scope of the
ikiwiki source covers the whole site to allow generated content (and
convenient wikilinks, etc) on all pages (such as news and release rss
feeds).

The job of migrating pages from the current MoinMoin wiki will be
undertaken mostly manually, per discussion at the summit - it's a good
opportunity to review their content anyway.  A number of key
pages will be translated across initially, to get the general setup
and stylesheet stuff into a decent state.

This work will take a little while, once it's progressed usefully far
this branch can be propagated back to nvm.web, and the webserver will
be switched to generate rendered .html at that time.  In the meantime,
mainline web changes can be prop'd here as usual.  Keeping wiki
changes in sync may prove harder, so the switch will happen once a
useful core of pages has been translated.