Remove include of <botan/keypair.h>. This header does not seem to be necessary with any possible version of botan because the only signatures it has ever provided are KeyPair::check_key, and whatever is in look_pk.h. Since neither file uses KeyPair::check_key, and both of them include look_pk.h already, the include is entirely extraneous.
This isn't really a big deal but it seems useful to remove unnecessary
dependencies, especially as with recent (>=1.8) botan it's possible to
disable keypair, and since it's not necessary, why rely on it being there?

(This is also relevant because in botan 1.9.4 and later, keypair.h might not
be visible to applications at all).