Start of get_output_color Lua hook being used by the colorizer
There is now a new Lua hook "get_output_color" with a default
implementation provided in std_hooks.lua to mimic the original
code. The hook maps a friendly "purpose name" (see 
colorizer::purpose and purpose_to_name) to a friendly
color name, which is then converted to the relevant ASCII escape
code.

As the colorizer now calls the get_output_color hook, it now
requires a lua_hooks instance in the constructor - and each of
the current callers currently pass in app.lua.

I have tried to keep the "const correctness" as mentioned in
HACKING - by making either the parameters const, as well as
the method const if possible.

NOTE: It's most likely that the default get_output_color
implementation could, and should, be cleaned up into a
more natural Lua style - instead of the if/elseif pattern.
Same goes for the color_to_code implementation as that uses
if/else if - although this might not be too bad as it's in 
C++.