database::get_var and var_exists now only ask for a single row
Previously they called get_vars which brought back the whole table
and like the FIXME said, it's not ideal but doesn't really matter
as theres only a handful of entries usually in db_vars.

Having said that, this patch removes a FIXME from the tree and has
the potential to improve performance ever so slightly.