Whitespace cleanup and reformatting
Most of the changes are due to misaligned code, or incorrect bracket
placement for GNU style. The following command line was used for the
formatting

astyle --style=gnu -s2 -p --align-pointer=middle --keep-one-line-blocks
--keep-one-line-statements --indent-namespaces
--max-instatement-indent=60 src/*.{cc,hh}

This minimizes the change of existing code formatting, whilst formatting it
according to the standards in HACKING, such as GNU and pointer placement.
The keep-one-line-blocks/statements options were used as to respect the
original code.

Most of the misaligned code is at the block level, usually by 1 space and
occured for if/else/for
eg.

if (condition)
  {
   // code was here
   ..
   // more code
  }

should have been

if (condition)
  {
    // code was here
    ..
    // more code
  }

However, another common case was argument misplacement

eg.

function_foo(p1, a_long_variable_name,
       p2, another_function(a, b, c));

should have been, according to most other code in the tree

function_foo(p1, a_long_variable_name,
             p2, another_function(a, b, c));

NOTE: Not all of the changes performed by astyle are to be kept.

1) The annotate template is now less readble than the original version, so
this will be reverted in a future commit.

2) struct/class now have an extra space at the start of the line before the
name

eg.
this change will make some struct/class's appear like

struct
 the_name

but should really be (to keep to the standard for functions)

struct
the_name

or possibly

struct the_name ?