Add get_real_branch_name hook to allow aliases in branch selectors
At the moment it's only used in branch selectors, but future commits will
allow it for any branch names from the user such as checkout and propagate.

Using the hook for netsync patterns would be useful but we'd probably have
to only call it for patterns without wildcards, or only allow wildcards
at the end; i.e. allowing aliases to be a prefix.

The default definition for this hook just returns branch_alias so
everything will be the same as before for people who don't want/need this hook.