Add 'filemodehint' parameter to 'temp_file' and 'write_to_temporary_file'
When writing files for external tools, such as loading the user's 3-way merge
tool to resolve a merge conflict, we want the contents to be exact; both for
presentation and the resulting file which will be used for the commit.

In other cases, like adding a 'comment' cert, we still work as before and
allow possible line-ending sanitisation.